Various small fixes, mostly memory leaks. Thanks to valgrind.
	* lib/ev-file-helpers.c: (ev_file_helpers_shutdown):
	* shell/ev-sidebar-links.c: (create_loading_model):
	* shell/ev-sidebar-thumbnails.c: (ev_sidebar_thumbnails_dispose):
	* shell/ev-view.c: (ev_view_set_document):
	* shell/ev-window.c: (ev_window_dispose):

	Various small fixes, mostly memory leaks. Thanks to valgrind.

+++ lib/ev-file-helpers.c
@@ -80,8 +80,9 @@ ev_file_helpers_init (void)
 
 void
 ev_file_helpers_shutdown (void)
-{		
-	rmdir (tmp_dir);
+{	
+	if (tmp_dir != NULL)	
+		rmdir (tmp_dir);
 
 	g_free (tmp_dir);
 	g_free (dot_dir);
